requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
mysqlclient==2.0.3
redis==3.5.3
celery==5.1.2
django_redis
requests
djangorestframework-simplejwt
python-dotenv
